New POISSON: 2 option (Gaussian sampling) and Flask Manual details
In this version we have included the possibility of sampling a galaxy
field using Gaussian noise instead of Poisson noise (POISSON: 2 instead
of 1). The standard deviation of the Gaussian noise is given by
square-root of the expected number of counts for a homogeneous universe
(and observed according to the selection function). We also included
negative density counting in both the POISSON: 0 and POISSON: 2, but
none of these are set to zero.

The manual was changed accordingly, and we seized the opportunity to
improve explanations regarding effects of certain keywords (such DIST:
GAUSSIAN and negative densities) and advices about checking certain
outputs to detect errors.