When sampling galaxy fields: fixed counting of negative values
Given that the masked regions in a healpix map can have negative
values (i.e. UNSEEN), this was making the counting of pixels with
negative number of galaxies wrong. We fixed this for the POISSON: 0 and
POISSON: 2 cases by only counting unmasked pixels. For the POISSON: 1
case, the counting is done on the density contrast map before the mask
is applied, so it reflects the full-sky fraction that had negative values,
and the UNSEEN does not affect the counting.