New option SELEC_SEPARABLE: 2
I changed the SelectionFunction class to allow a new kind of selection function separability:
in option 2, each tracer have a different angular selection function, instead of sharing the
same one (as is done in option 1). The code has to be tested and the changes documented. The
next commit should do that.