Generalized prepClassInput.py to UCLCl output
UCLCl output has a similar format to CLASS', but with fewer header lines,
it may output 'temp' columns and it starts Cl indexes at zero. I've made
prepClassInput.py be able to handle any number of headers (all starting
with hash and assuming the column names are the last one), to handle 'temp'
(only tested with temp-only Cl table, but should work for combinations with
'dens' and 'lens') and to change bin numbering starting from zero to one.

I have noticed that there are operations made for columns 'phi', but there
is no handling of this header.

Also, the testing, documenting and commiting of the changes mentioned in
the previous commit are stil to be done.