The Cl truncation at LMAX is now an option CROP_CL
In commit 62e0a9, I changed the code so it would always use in the transf.
from lognormal to Gaussian Cls the multipoles up to LMAX set in LRANGE.
Now this approach and the previous one (to use in the transf. all multipoles
in the input Cls) are chosen with the CROP_CL keyword. example.config set to
default choice and manual updated accordingly.