Single table in CL_PREFIX working and new script mergeCls.py
In this version the CL_PREFIX keywork also accepts a single file terminated
by '.dat'. If such file is passed to CL_PREFIX, all required Cls will be read
from it. Its first column should be L, and then the Cls in any order. It is
required that the file contains a header starting with hash '#' and with the
name of the columns. The first name is ignored (since it is assumed to be L)
and the remaining ones should contain the FLASK naming pattern f<F1>z<Z1>f<f2>z<z2>.
This header is used to assign each Cl to its proper place in the covariance
matrices. Previous option for CL_PREFIX (a prefix for individual Cl files) still
works if CL_PREFIX is sert to something that does not end in '.dat'.

We also created the mergeCls.py script, which takes a series of individual Cl files
and place them in a single file.

TODO:
-- Document this changes in the manual;
-- Make Cl outputs to follow the same standard as used in CL_PREFIX.