Optional single Cl table for input, smooth and regularized outputs.
I finished implementing the possibility to pass a single .dat file to CL_PREFIX
containing all required Cls in a simulation. This is interesting in the case where
the number of fields to be simulated (and, consequently, the number of Cls) is
very large (it is just an organizational change). I also included the "single Cl
table file" option for SMOOTH_CL_PREFIX and REG_CL_PREFIX. For the former, the
output order is the same as the input order. For the latter, the output is FLASK's
standard (i.e. follows the order in FIELDS_INFO file and with larger field and
redshift numbers in the last Cl field slot). I have tested the code with a few
field combinations and file formats and configurations, and it seems to be working
fine.

In the previous commit, I have also fixed a bug related to std::complex variables.
It seems that in a recent C++ update (I think C++11 or something) the way to assign
numbers to the real and imaginary parts of complex numbers changed. I have changed
the FLASK code accordingly (e.g. gamma.real() = g1; to gamma.real(g1); ). This change
 only affects the computation of ellipticities. I have checked that the new code
compiles even in the older std::complex definition, but I have not tested the impact
on the ellipticities. I guess if it compiles, it works properly.