Small bug fixes found by Nicolas
The declaration of the warning function was contradictory, and this was causing problems
in some versions of gcc. This was fixed by removing redundant declarations.
Also, Healpix (3.50) uses pocketfft instead of fftpack library. I changed the Makefile
to check the Healpix version and use the correct library. Only tested for Healpix 3.50.
Everything is working.